Issues/12 (#66)
A new attribute, `clusterLetters` was added to valued options, in order
to support the so called "short-options" style. It was grouped with the
`envVar` attribute under a `WithMisc` type.

A new property, `shortStyle`, was added to the parser configuration, to
indicate whether the first command-line argument is expected to be an
option cluster. This is to be used in conjunction with the above.

The Parser page was updated to document the `shortStyle` property. The
Options page was updated to document the `clusterLetters` attribute
under a new section called `Miscellaneous attributes`, together with the
`envVar` attribute.

Closes #12