Release 0.3.0 (#67)
* Enter prerelease mode

* Issues/9 (#33)

A new field called `progName` was added to `ParseConfig`, that allows
program authors to specify a custom program name to be used to update
the process title.

The Parser page of the documentation was updated to reflect the above
change, as well as add a new section to explain how the process title is
updated by the parser.

Closes #9

* Issues/16 (#41)

A new guide page called "Help" was added to the documentation. It
describes how to implement different help mechanisms for a CLI
application.

A new command, `help`, was added to the demo example, to demonstrate one
of the mechanisms explained in the above guide.

Closes #16

* Issues/24 (#42)

A new `envVar` attribute was added to options that have a value, to read
the value from an environment variable when it is not specified on the
command-line.

The `requires` attribute was removed from special options, since these
options do not have values (they only throw a message, so there's no
point in checking their requirements).

The `default` attribute was added to executing options, with an
`unknown` value type. This allows program authors to define any value as
the default for function and command options.

A new section called "Value attributes" was added to the Options page.
It includes the atrtibutes common to all valued option types, but with
one caveat: the `envVar` attribute is not available for executing options.

Closes #24

* Version Packages (dev) (#34)

* [release] Releasing 2 package(s)

- tsargp@0.3.0-dev.0
- @trulysimple/tsargp-docs@0.2.0-dev.0

* Version Packages (dev)

* Issues/44 (#45)

The `tsargp` package was refactored to remove unneeded exports, which
reduced the footprint of the minified module by 0.36KB.

Some bindings were marked with the `@internal` tag, in order to strip
them from the emitted declaration files. This also helped reduce the
overall package size.

Control sequence enumerations have been unified into a single
`ControlSequence`, and related functions for creating sequences have
been unified into a single `seq` function. This did not affect the
`Style` type nor the `style` function: they were preserved as
specializations.

**Developer note**
Henceforth, marking a type with the `@internal` tag will strip it from
declaration files. This should be used with care: only use it for types
that are not referenced by any other exported type.

Closes #44

* Issues/11 (#48)

New classes were created for the messages that the parser may produce:

- `TerminalMessage` - base class for other message types
- `WarnMessage` - represents a warning message
- `CompletionMessage` - the list of completion words
- `VersionMessage` - the version (currently, just a string)

New methods were added to `ArgumentParser` to ease handling of messages:

- `doParse` - the most flexible method to parse arguments
- `tryParse` - convenience method that catches errors before returning

A new enumerator, `deprecatedOption`, was included in `ErrorItem` to
represent a warning of deprecated options.

The following types were removed, as they are no longer needed:

- `OpaqueArgumentParser`
- `CastToOptionValues`

Changes in the documentation:

- a new section called "Terminal messages" was added to the Styles page
to describe the classes listed above
- the Parser page was updated to document the new parsing methods
- the `deprecatedOption` warning was added to the Validator page

Closes #11

* Ignore pre.json in cspell

* Enable hidden directories in cspell

* Version Packages (dev) (#46)

* [release] Releasing 2 package(s)

- @trulysimple/tsargp-docs@0.2.0-dev.1
- tsargp@0.3.0-dev.1

* Version Packages (dev)

* Refactor/10 (#50)

The unit tests were refactored into subdirectories, one for each library
component.

* Issues/10 (#49)

A new attribute,`requiredIf`, was added to valued options, to indicate
the option's conditional requirements. An equally named enumerator was
added to `HelpItem`, in order to print this attribute in the help message.

Both the parser and the formatter were updated to handle this attribute,
and many test cases were added in order to cover the new functionality.

The Options page was updated to document the new attribute, with an
example that illustrates the difference between it and the traditional
`requires`.

Closes #10

* Version Packages (dev) (#51)

* [release] Releasing 2 package(s)

- @trulysimple/tsargp-docs@0.2.0-dev.2
- tsargp@0.3.0-dev.2

* Version Packages (dev)

* Refactor/19 (#52)

Implemented recursive validation of command options.

The Formatter page in the documentation was refactored to flatten out
the "Help format" section.

* Issues/19 (#53)

Added the `formatFull` method to the help formatter, that implements
sections in the help message. The full help can be configured with the
new `HelpConfig` properties: `sections`, `styles` amd `misc`.

Some attributes were removed from the help option, as they are not
needed anymore.

The Formatter page was updated to document the new `formatFull` method.

Closes #19

* Issues/54 (#57)

Added missing source links in the API documentation generated with
TypeDoc.

Closes #54

* Issues/55 (#58)

Added `@cspell/eslint-plugin` as a dev dependency and configured
`eslint.config.js` to check files with it.

Closes #55

* Refactor/14 (#59)

Refactored the `formatFull` method into `formatSections`.

* Issues/47 (#60)

Added a help format configuration named `align`, that can be used to
select the alignment of option names in the help message.

Closes #47

* Version Packages (dev) (#56)

* [release] Releasing 2 package(s)

- @trulysimple/tsargp-docs@0.2.0-dev.3
- tsargp@0.3.0-dev.3

* Version Packages (dev)

* Refactor/13 (#62)

Fixed a bug with example array values in the help message parameter
column.

Refactored unit tests for text wrapping into a new file,
`styles.wrapping.spec.ts`.

* Issues/13 (#61)

New properties, `param` and `descr`, were added to the text alignment
settings of `HelpConfig`. A new property, `rightAlign`, was added to the
`TerminalString` class, that makes the `wrapToWidth` method align text
to the terminal's right boundary.

The Formatter page was updated to document the new alignment settings.
A new section was added to the Styles page, to document the new
right-alignment feature of the wrapping procedure.

Closes #13

* Refactor/12 (#64)

A new type called `WithColumn` was added. It is used in the `HelpConfig`
object for configuring the help columns.

The wrapping procedure was improved to emit all control sequences when
`emitStyles` is true.

* Refactor/12 (#65)

An ellipsis ('...') was added to the parameter of variadic array options
in the help message.

* Issues/12 (#66)

A new attribute, `clusterLetters` was added to valued options, in order
to support the so called "short-options" style. It was grouped with the
`envVar` attribute under a `WithMisc` type.

A new property, `shortStyle`, was added to the parser configuration, to
indicate whether the first command-line argument is expected to be an
option cluster. This is to be used in conjunction with the above.

The Parser page was updated to document the `shortStyle` property. The
Options page was updated to document the `clusterLetters` attribute
under a new section called `Miscellaneous attributes`, together with the
`envVar` attribute.

Closes #12

* Version Packages (dev) (#63)

* [release] Releasing 2 package(s)

- tsargp@0.3.0-dev.4
- @trulysimple/tsargp-docs@0.2.0-dev.4

* Version Packages (dev)

* Resolve review comments

* Exit prerelease mode

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>