Issues/79 (#84)
The parser was refactored to more easily handle asynchronous callbacks.
As a consequence, it no longer supports a synchronous parsing method. On
the other hand, this change brought the footprint down to 31KB.

A new kind of validation, `invalidRequiredValue`, was implemented for
the verification of required option values in the case of options that
are either always required or have a default value.

A whole section about asynchronous parsing was removed from the Parser
page, according to changes made in the code. The Validator page was
updated to document the new validation mentioned above.