Issues/90 (#92)
The parsing callbacks were refactored to accept a single parameter of
type `ParseInfo`, which contains information about the current argument
sequence in the parsing loop.

The parsing loop was refactored to accumulate a sequence of arguments
that are then passed to the parsing callbacks as the option
parameter(s).

As a consequence of this change, name suggestions on parse errors are no
longer supported. Accordingly, the `parseError` enumerator was removed
from `ErrorItem`.

The Parser page was updated to document the new `ParseInfo` type, along
with information regarding the parsing callbacks. The Options page was
updated to document changes made to the parsing callbacks. The
`parseError` enumerator was removed from the Validator page.

Closes #90