Refactor/83 (#94)
A new attribute, `saveMessage`, was added to help and version options to
indicate that the message should be saved in the option value instead of
being thrown.

All attributes of the function and command options were made optional,
in order to be consistent with the other option types.

A new property, `isComp`, was added to the parameter of the function
callback, that allows the callback to check whether any of the remaining
command-line arguments is a word to be completed.

The "Option types" section of the Options page was refactored to explain
in more detail the information presented in the table. The `saveMessage`
attribute was documented in both the help and version options' sections.