Refactor/83 (#95)
Added new attributes for the boolean option: `truthNames`,
`falsityNames` and `caseSensitive`. They can be used to configure how
option parameters are converted to boolean.

Added a new enumerator called `invalidBooleanParameter` to `ErrorItem`.
It is used by the parser to throw an error when a conversion from string
to boolean fails.

Some attributes of the `WithParam` type were moved to a new type called
`WithKnownValue`, in anticipation for the feature that will be
implemented in #83.

Some sections of the Options page were updated to reflect the changes
made in the code.