Releases/0.5.0 (#126)
* Enter pre-release mode

* Update packages

* Issues/104 (#105)

Added the `useNested` attribute to the help option, which allows a
comand-line argument following the option to be used as the name of a
nested command for which the help message should be assembled. This
feature is _not_ mutually exclusive with `useFilter`.

Documentation changes:
- Updated Xterm.js to 5.5.0.
- Added documentation for the `useNested` attribute of help options in
the Options page.
- Updated the Help guide to discourage the use of help commands in favor
of the new attribute.

Closes #104

* Issues/107 (#109)

Updated the documentation in several places to fix typos and tidy things
up.

Closes #107

* Refactor/18 (#111)

Updated the code and documentation to consider the boolean's truth and
falsity names wherever the `enums` attribute is used.

The validator tests were refeactored to more closely match how the
documentation of the Validator page is organized.

* Issues/18 (#110)

Replaced the ANSI escape code `\x9b` by the more traditional `\x1b[`, in
order to support terminals that do not recognize the former.

Fixed the `getArgs` procedure to support word completion in PowerShell.

Updated the Introduction page to document how word completion can be
enabled for PowerShell.

Closes #18

* Issues/112 (#113)

Refactored the parser to ignore command-line arguments coming after the
completion index.

Removed the `isComp` utility function and the `WithIsComp` type, as they
are no longer needed.

Removed documentation of the `isComp` property from the Parser page.

Closes #112

* Issues/114 (#115)

Added reading of command line and completion index from the `BUFFER` and
`CURSOR` environment variables, respectively.

Updated the Introduction page to document how to enable word completion
in Z shell.

Closes #114

* Fix CSpell config

* Issues/116 (#117)

Patched the `getArgs` utility function to handle characters escaped with
a backslash.

Updated the Introduction page to document how word completion can be
enabled in Fish shell.

Closes #116

* Issues/118 (#119)

Updated the parsing methods to accept the `BUFFER` environment variable
without the `CURSOR` variable.

Updated the Introduction page to document how to enable word completion
for Nushell.

Closes #118

* Issues/120 (#121)

Replaced the `shortStyle` configuration flag by the more generic
`clusterPrefix`, which allows cluster arguments anywhere in the command
line.

Updated the Parser page to document the new flag, as well as the Options
page to document the new behavior of cluster arguments.

Closes #120

* Issues/122 (#123)

Implemented support for inline cluster parameters (this is not to be
confused with inline parameters of normal arguments).

Updated the Parser and Options pages to document the new syntax for
inline cluster parameters.

Closes #122

* Version Packages (dev) (#106)

* [release] Releasing 2 package(s)

- tsargp@0.5.0-dev.0
- @trulysimple/tsargp-docs@0.4.0-dev.0

* Issues/124 (#125)

Cleaned up the code. This brought the footprint down to 32.2KB.

Closes #124

* Exit pre-release mode

* Resolve review comments

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>