Issues/132 (#133)
Implemented handling of asynchronous `OptionsCallback` for the command
option. This allows loading option definitions from a different module
using a dynamic `import`. This required the `validate` method to return
a promise.

Updated the Options page to document the support for this new feature in
the "Command option" section. Also updated the Validator page to
document the new asynchronous behavior of the `validate` procedure.

Closes #132