Releases/0.6.0 (#146)
* Enter pre-release mode

* Issues/132 (#133)

Implemented handling of asynchronous `OptionsCallback` for the command
option. This allows loading option definitions from a different module
using a dynamic `import`. This required the `validate` method to return
a promise.

Updated the Options page to document the support for this new feature in
the "Command option" section. Also updated the Validator page to
document the new asynchronous behavior of the `validate` procedure.

Closes #132

* Issues/136 (#137)

Extended the connective words to include array element separators and
string quoting. The new enumerators are: `optionAlt`, `optionSep`,
`stringSep`, `numberSep` and `stringQuote`.

Renamed the `ConnectiveWords` enumeration to singular, and created
another type for the object holding the words themselves.

Updated the Validator page to document the new connective words.

Closes #136

* Issues/139 (#140)

Added a `JsonFormatter` class that handles formatting of help messages
in JSON format, as well as a `HelpFormat` type for the available help
formats.

The previous `HelpFormatter` class was renamed to `AnsiFormatter`, while
the previous name is now used for the interface implemented by these
classes.

The `format` attribute of help options was renamed to `config`, while
the previous name is now used to indicate the help format.

Updated the Options page to document the `config` and `format`
attributes of help options. Updated Formatter page to document the new
organization of formatter classes.

Closes #139

* Issues/141 (#143)

Added the `CsvFormatter` class to handle formatting of help messages in
CSV format.

Updated the Formatter page to document the new formatter class.

Closes #141

* Issues/142 (#144)

Added help items for the attributes of a help option: `useNested`,
`useFormat` and `useFilter`.

Updated the Formatter page to document the new help items.

Closes #142

* Issues/135 (#145)

Added the `MdFormatter` class to format help messages in Markdown
format.

Updated the Formatter page to document the new formatter class.

Closes #135

* Clean up the code

* Exit pre-release mode

* Resolve review comments

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>