Issues/154 (#155)
Added an ellipsis `...` as parameter suffix for example values of variadic
options, as well as an equals sign `=` as prefix for options that require
inline parameters. Updated the Formatter page to document the new syntax.

Altered the parser to disallow inline parameters of variadic options,
and updated the Options and Parser pages accordingly.

Added a new validation error, `invalidInlineConstraint`, raised when an
inline parameter constraint is declared in a variadic option. Updated
the Validator page accodingly.

Closes #154