Releases/0.7.0 (#167)
* Eneter pre-release mode

* Update dev dependencies

* Update documentation

* Issues/149 (#150)

Added the `noInline` attribute for non-niladic options, to disallow
inline parameters.

Updated the Options page to document the new attribute for non-niladic
options.

Closes #149

* Issues/152 (#153)

Renamed the `noInline` attribute to `inline` and added support for a new
value `always`, which indicates that the option _requires_ inline
parameters. Updated the Options page accordingly.

Added the help item `inline` to report if an option disallows or
requires inline parameters. Updated the Formatter page accordingly.

Closes #152

* Issues/154 (#155)

Added an ellipsis `...` as parameter suffix for example values of variadic
options, as well as an equals sign `=` as prefix for options that require
inline parameters. Updated the Formatter page to document the new syntax.

Altered the parser to disallow inline parameters of variadic options,
and updated the Options and Parser pages accordingly.

Added a new validation error, `invalidInlineConstraint`, raised when an
inline parameter constraint is declared in a variadic option. Updated
the Validator page accodingly.

Closes #154

* Issues/158 (#159)

Added a new value `merge` for the `align` property of help columns,
which instructs the formatter to merge the contents of the column with
the previous one. The formatter page was updated accordingly.

Closes #158

* Issues/160 (#161)

Updated the Options and Validator pages to add more details about option
requirements.

Closes #160

* Issues/157 (#163)

Added a `requires` attribute in the usage section of help messages, to
support option dependencies. This is related to option requirements, but
somewhat different. Updated the Formatter page to document this.

Closes #157

* Issues/156 (#164)

Updated the Home page to show the help message of the demo example.

Closes #156

* Issues/165 (#166)

The `example` value is now enclosed in square brackets in the help
message when the option has a fallback value, just as is done with
parameter names and option types.

Added the `.js` extension to all library imports, to fix import issues
in the `lib` export path.

Updated the Home page with the new help message of the demo example.

Closes #165

- Fix `this` in callbacks
- Add visitRequirements
- Clean up the code

* Exit pre-release mode

* Update dev dependencies

* Resolve review comments

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>